[SW-1483] Introduce the new package and remove the old ones
