+++ ml/src/main/scala/ai/h2o/sparkling/ml/algos/H2OXGBoost.scala
@@ -14,17 +14,17 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-package org.apache.spark.ml.h2o.algos
+package ai.h2o.sparkling.ml.algos
 
 import ai.h2o.sparkling.macros.DeprecatedMethod
-import ai.h2o.sparkling.ml.algos.H2OSupervisedAlgorithm
 import ai.h2o.sparkling.ml.params.H2OAlgoParamsHelper._
 import ai.h2o.sparkling.ml.params.H2OAlgoSupervisedParams
+import ai.h2o.sparkling.ml.utils.H2OParamsReadable
 import hex.schemas.XGBoostV3.XGBoostParametersV3
 import hex.tree.xgboost.XGBoostModel.XGBoostParameters
 import hex.tree.xgboost.XGBoostModel.XGBoostParameters._
 import hex.tree.xgboost.{XGBoost, XGBoostModel}
-import org.apache.spark.ml.util.{DefaultParamsReader, Identifiable}
+import org.apache.spark.ml.util.Identifiable
 
 /**
   * H2O XGBoost algorithm exposed via Spark ML pipelines.
@@ -34,7 +34,7 @@ class H2OXGBoost(override val uid: String) extends H2OSupervisedAlgorithm[XGBoos
   def this() = this(Identifiable.randomUID("xgboost"))
 }
 
-object H2OXGBoost extends DefaultParamsReader[py_sparkling.ml.algos.H2OXGBoost]
+object H2OXGBoost extends H2OParamsReadable[H2OXGBoost]
 
 
 /**
