+++ py/ai/h2o/sparkling/ml/algos/__init__.py
@@ -15,3 +15,9 @@
 # limitations under the License.
 #
 from ai.h2o.sparkling.ml.algos.H2OKmeans import H2OKMeans
+from ai.h2o.sparkling.ml.algos.H2ODeepLearning import H2ODeepLearning
+from ai.h2o.sparkling.ml.algos.H2OXGBoost import H2OXGBoost
+from ai.h2o.sparkling.ml.algos.H2OGBM import H2OGBM
+from ai.h2o.sparkling.ml.algos.H2OGLM import H2OGLM
+from ai.h2o.sparkling.ml.algos.H2OGridSearch import H2OGridSearch
+from ai.h2o.sparkling.ml.algos.H2OAutoML import H2OAutoML
