+++ py/ai/h2o/sparkling/ml/features/ColumnPruner.py
@@ -1,3 +1,20 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
 from pyspark import keyword_only
 from pyspark.ml.param.shared import *
 from pyspark.ml.util import JavaMLReadable, JavaMLWritable
@@ -16,7 +33,7 @@ class ColumnPruner(JavaTransformer, JavaMLReadable, JavaMLWritable):
     def __init__(self, keep=False, columns=[]):
         Initializer.load_sparkling_jar()
         super(ColumnPruner, self).__init__()
-        self._java_obj = self._new_java_obj("py_sparkling.ml.features.ColumnPruner", self.uid)
+        self._java_obj = self._new_java_obj("ai.h2o.sparkling.ml.features.ColumnPruner", self.uid)
         self._setDefault(keep=False, columns=[])
         kwargs = get_input_kwargs(self)
         self.setParams(**kwargs)
